% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blrm.r
\name{as.data.frame.Ocens}
\alias{as.data.frame.Ocens}
\title{as.data.frame.Ocens}
\usage{
\method{as.data.frame}{Ocens}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{an \code{Ocens} object}

\item{row.names}{optional vector of row names}

\item{optional}{set to \code{TRUE} if needed}

\item{...}{ignored}
}
\value{
data frame containing a 2-column integer matrix with attributes
}
\description{
Convert \code{Ocens} Object to Data Frame to Facilitate Subset
}
\details{
Converts an \code{Ocens} object to a data frame so that subsetting will preserve all needed attributes
}
\author{
Frank Harrell
}
