% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanMisc.r
\name{getParamCoef}
\alias{getParamCoef}
\title{getParamCoef}
\usage{
getParamCoef(
  fit,
  posterior.summary = c("mean", "median", "mode"),
  what = c("both", "betas", "taus")
)
}
\arguments{
\item{fit}{a Bayesian model fit from \code{rmsb}}

\item{posterior.summary}{which summary statistic (Bayesian point estimate) to fetch}

\item{what}{specifies which coefficients to include.  Default is all.  Specify \code{what="betas"} to include only intercepts and betas if the model is a partial proportional odds model (i.e.,, exclude the tau parameters).  Specify \code{what="taus"} to include only the tau parameters.}
}
\value{
vector of regression coefficients
}
\description{
Get a Bayesian Parameter Vector Summary
}
\details{
Retrieves posterior mean, median, or mode (if available)
}
\author{
Frank Harrell
}
