% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blrm.r
\name{Ocens}
\alias{Ocens}
\title{Ocens}
\usage{
Ocens(a, b = a)
}
\arguments{
\item{a}{vector representing a \code{factor}, numeric, or alphabetically ordered character strings}

\item{b}{like \code{a}.  If omitted, it copies \code{a}, representing nothing but uncensored values}
}
\value{
a 2-column integer matrix of class \code{"Ocens"} with an attribute \code{levels} (ordered).  When the original variables were \code{factor}s, these are factor levels, otherwise are numerically or alphabetically sorted distinct (over \code{a} and \code{b} combined) values.  When the variables are not factors and are numeric, another attribute \code{median} is also returned.  This is the median of the uncensored values.  When the variables are factor or character, the median of the integer versions of variables for uncensored observations is returned as attribute \code{mid}.  A final attribute \code{freq} is the vector of frequencies of occurrences of all uncensored values.  \code{freq} aligns with \code{levels}.
}
\description{
Censored Ordinal Variable
}
\details{
Creates a 2-column integer matrix that handles left- right- and interval-censored ordinal or continuous values for use in \code{\link[=blrm]{blrm()}}.  A pair of values \verb{[a, b]} represents an interval-censored value known to be in the interval \verb{[a, b]} inclusive of \code{a} and \code{b}.  It is assumed that all distinct values are observed as uncensored for at least one observation.  When both input variables are \code{factor}s it is assume that the one with the higher number of levels is the one that correctly specifies the order of levels, and that the other variable does not contain any additional levels.  If the variables are not \code{factor}s it is assumed their original values provide the orderings.  Since all values that form the left or right endpoints of an interval censored value must be represented in the data, a left-censored point is is coded as \code{a=1} and a right-censored point is coded as \code{b} equal to the maximum observed value.  If the maximum observed value is not really the maximum possible value, everything still works except that predictions involving values above the highest observed value cannot be made.  As with most censored-data methods, \code{\link[=blrm]{blrm()}} assumes that censoring is independent of the response variable values that would have been measured had censoring not occurred.
}
\author{
Frank Harrell
}
