% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soprobMarkovOrdPost.r
\name{soprobMarkovOrdPost}
\alias{soprobMarkovOrdPost}
\title{soprobMarkovOrdPost}
\usage{
soprobMarkovOrdPost(
  object,
  data,
  times,
  ylevels,
  tvarname = "time",
  pvarname = "yprev",
  gap = NULL
)
}
\arguments{
\item{times}{vector of measurement times}

\item{y}{a vector of possible y values in order (numeric, character, factor)}

\item{initial}{initial value of \code{y} (baseline state; numeric, character, factr)}

\item{absorb}{vector of absorbing states, a subset of \code{y}.  The default is no absorbing states. (numeric, character, factor)}

\item{intercepts}{vector of intercepts in the proportional odds model, with length one less than the length of \code{y}}

\item{...}{additional arguments to pass to \code{g} such as covariate settings}
}
\value{
matrix with rows corresponding to times and columns corresponding to states, with values equal to exact state occupancy probabilities
}
\description{
State Occupancy Probabilities for First-Order Markov Ordinal Model over Posterior Draws
}
\seealso{
\url{https://hbiostat.org/R/Hmisc/markov/}
}
\author{
Frank Harrell
}
